+++ packages/utils/CHANGELOG.md
@@ -3,11 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-# 2.2.0 (2022-09-25)
-
+# [2.3.0](http://jopau-react/compare/@jopau-react/utils@2.2.0...@jopau-react/utils@2.3.0) (2022-09-25)
 
 ### Features
 
-* **utils:** add function utils ([85e9a8c](http://jopau-react/commits/85e9a8ce7d8630c971a3755b41a1c1e64d158ebb))
-* **utils:** create monorepo package ([fb253ea](http://jopau-react/commits/fb253ea5c9a367bf71d0a2f95bdaf6f46a043503))
-* **utils:** update readme ([3e004d3](http://jopau-react/commits/3e004d3c836f53f0cf7ada82cbe581982d34729a))
+- **utils:** add some, size and includes collection methods ([33d0ae2](http://jopau-react/commits/33d0ae27b515689e81329c1404c251b9974ec41a))
+
+# [2.3.0](http://jopau-react/compare/@jopau-react/utils@2.2.0...@jopau-react/utils@2.3.0) (2022-09-25)
+
+### Features
+
+- **utils:** add some, size and includes collection methods ([589e9af](http://jopau-react/commits/589e9af6e6082470c5926eaff944caca38e9fb15))
+
+# 2.2.0 (2022-09-25)
+
+### Features
+
+- **utils:** add function utils ([85e9a8c](http://jopau-react/commits/85e9a8ce7d8630c971a3755b41a1c1e64d158ebb))
+- **utils:** create monorepo package ([fb253ea](http://jopau-react/commits/fb253ea5c9a367bf71d0a2f95bdaf6f46a043503))
+- **utils:** update readme ([3e004d3](http://jopau-react/commits/3e004d3c836f53f0cf7ada82cbe581982d34729a))